+++ libraries/TFormat/TScaler.cxx
@@ -284,8 +284,9 @@ TH1D* TScaler::Draw(UInt_t address, size_t index, Option_t* option)
       if(fHist[address] == nullptr) {
          int nofBins    = fPPG->GetCycleLength() / GetTimePeriod(address);
          fHist[address] = new TH1D(Form("TScalerHist_%04x", address),
-                                   Form("scaler %d vs time in cycle for address 0x%04x; time in cycle [ms]; counts/%.0f ms", static_cast<int>(index), address, fPPG->GetCycleLength() / 1e5 / nofBins),
-                                   nofBins, 0., fPPG->GetCycleLength() / 1e5);
+                                   Form("scaler %d vs time in cycle for address 0x%04x; time in cycle [ms]; counts/%.0f ms",
+                                        static_cast<int>(index), address, static_cast<double>(fPPG->GetCycleLength()) / 1e5 / nofBins),
+                                   nofBins, 0., static_cast<double>(fPPG->GetCycleLength()) / 1e5);
          // fHist[address]->ResetBit(kMustCleanup);
       }
       // we have to skip the first data point in case this is a sub-run
@@ -297,7 +298,7 @@ TH1D* TScaler::Draw(UInt_t address, size_t index, Option_t* option)
             // fill the difference between the current and the next scaler (if we found a previous value and that one is
             // smaller than the current one)
             if(previousValue != 0 && previousValue < fScalerData->GetScaler(index)) {
-               fHist[address]->Fill(fPPG->GetTimeInCycle(fScalerData->GetTimeStamp()) / 1e6,
+               fHist[address]->Fill(static_cast<double>(fPPG->GetTimeInCycle(fScalerData->GetTimeStamp())) / 1e6,
                                     fScalerData->GetScaler(index) - previousValue);
             }
             previousValue = fScalerData->GetScaler(index);
@@ -354,8 +355,9 @@ TH1D* TScaler::Draw(UInt_t lowAddress, UInt_t highAddress, size_t index, Option_
          int nofBins = fPPG->GetCycleLength() / GetTimePeriod(lowAddress);   // the time period should be the same for all channels
          fHistRange[std::make_pair(lowAddress, highAddress)] =
             new TH1D(Form("TScalerHist_%04x_%04x", lowAddress, highAddress),
-                     Form("scaler %d vs time in cycle for address 0x%04x - 0x%04x; time in cycle [ms]; counts/%.0f ms", static_cast<int>(index), lowAddress, highAddress, fPPG->GetCycleLength() / 1e5 / nofBins),
-                     nofBins, 0., fPPG->GetCycleLength() / 1e5);
+                     Form("scaler %d vs time in cycle for address 0x%04x - 0x%04x; time in cycle [ms]; counts/%.0f ms",
+                          static_cast<int>(index), lowAddress, highAddress, static_cast<double>(fPPG->GetCycleLength()) / 1e5 / nofBins),
+                     nofBins, 0., static_cast<double>(fPPG->GetCycleLength()) / 1e5);
          // fHistRange[std::make_pair(lowAddress, highAddress)]->ResetBit(kMustCleanup);
          // we have to skip the first data point in case this is a sub-run
          // loop over the remaining scaler data for this address
@@ -369,7 +371,7 @@ TH1D* TScaler::Draw(UInt_t lowAddress, UInt_t highAddress, size_t index, Option_
                if(previousValue[fScalerData->GetAddress()] != 0 &&
                   previousValue[fScalerData->GetAddress()] < fScalerData->GetScaler(index)) {
                   fHistRange[std::make_pair(lowAddress, highAddress)]->Fill(
-                     fPPG->GetTimeInCycle(fScalerData->GetTimeStamp()) / 1e6,
+                     static_cast<double>(fPPG->GetTimeInCycle(fScalerData->GetTimeStamp())) / 1e6,
                      fScalerData->GetScaler(index) - previousValue[fScalerData->GetAddress()]);
                }
                previousValue[fScalerData->GetAddress()] = fScalerData->GetScaler(index);
@@ -401,8 +403,8 @@ TH1D* TScaler::Draw(UInt_t lowAddress, UInt_t highAddress, size_t index, Option_
          fHist[address] =
             new TH1D(Form("TScalerHist_%04x", address),
                      Form("scaler %d vs time in cycle for address 0x%04x; time in cycle [ms]; counts/%.0f ms",
-                          static_cast<int>(index), address, fPPG->GetCycleLength() / 1e5 / nofBins),
-                     nofBins, 0., fPPG->GetCycleLength() / 1e5);
+                          static_cast<int>(index), address, static_cast<double>(fPPG->GetCycleLength()) / 1e5 / nofBins),
+                     nofBins, 0., static_cast<double>(fPPG->GetCycleLength()) / 1e5);
       } else {
          // if the histogram already exist, reset it
          fHist[address]->Reset();
@@ -419,7 +421,7 @@ TH1D* TScaler::Draw(UInt_t lowAddress, UInt_t highAddress, size_t index, Option_
          // smaller than the current one)
          if(previousValue[fScalerData->GetAddress()] != 0 &&
             previousValue[fScalerData->GetAddress()] < fScalerData->GetScaler(index)) {
-            fHist[fScalerData->GetAddress()]->Fill(fPPG->GetTimeInCycle(fScalerData->GetTimeStamp()) / 1e6,
+            fHist[fScalerData->GetAddress()]->Fill(static_cast<double>(fPPG->GetTimeInCycle(fScalerData->GetTimeStamp())) / 1e6,
                                                    fScalerData->GetScaler(index) -
                                                       previousValue[fScalerData->GetAddress()]);
          }
@@ -455,7 +457,7 @@ TH1D* TScaler::DrawRawTimes(UInt_t address, Double_t lowtime, Double_t hightime,
 
    TString opt = option;
    opt.ToLower();
-   int nofBins = std::abs(static_cast<int>(1e9 * (hightime - lowtime) / GetTimePeriod(address)));
+   int nofBins = std::abs(static_cast<int>(1e9 * (hightime - lowtime) / static_cast<double>(GetTimePeriod(address))));
    std::cout << nofBins << "nofbins" << std::endl;
    // This scHist could be leaky as the outside user has ownership of it.
    auto* scHist =
@@ -471,7 +473,7 @@ TH1D* TScaler::DrawRawTimes(UInt_t address, Double_t lowtime, Double_t hightime,
          // fill the difference between the current and the next scaler (if we found a previous value and that one is
          // smaller than the current one)
          if(previousValue != 0 && previousValue < fScalerData->GetScaler(index)) {
-            scHist->Fill(fScalerData->GetTimeStamp() / 1e9, fScalerData->GetScaler(index) - previousValue);
+            scHist->Fill(static_cast<double>(fScalerData->GetTimeStamp()) / 1e9, fScalerData->GetScaler(index) - previousValue);
          }
          previousValue = fScalerData->GetScaler(index);
       }
